requirements.txt
# Web Framework
fastapi>=0.100.0
uvicorn>=0.20.0
python-dotenv>=1.0.0
requests>=2.31.0

# Email Processing
imap-tools>=1.0.0

# AI & LangChain Ecosystem
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.10
langchain-openai>=0.0.5
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.1

# Vector Database & Embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# Document Loaders
pypdf>=4.0.0
tiktoken>=0.5.0